[Update] MetaWeatherRepository
data 에서 domain 에 있는 MetaWeatherRepository 를 구현했습니다.
MetaWeatherRepository 는 MetaWeather API 의 데이터를 가져오고, 그 데이터를 domain 에 정의된 entity 로 가공해주는 역할을 합니다.

+++ domain/src/main/java/com/meta/weather/domain/entity/WeatherEntity.kt
@@ -4,5 +4,5 @@ class WeatherEntity(
     val state: String,
     val icon: String,
     val temp: Float,
-    val humidity: Float
+    val humidity: Int
 )
